+++ core/testsuite/src/main/java/com/blazebit/persistence/testsuite/AbstractCoreTest.java
@@ -140,18 +140,11 @@ public abstract class AbstractCoreTest extends AbstractPersistenceTest {
         return jpaProvider.escapeCharacter(character);
     }
 
-    protected String likePattern(String expression) {
-        Character defaultEscapeCharacter = dbmsDialect.getDefaultEscapeCharacter();
-        if (jpaProvider.supportsLikePatternEscape() || defaultEscapeCharacter == null) {
-            return expression;
+    protected String noEscape() {
+        if (!jpaProvider.supportsLikePatternEscape() && dbmsDialect.getDefaultEscapeCharacter() != null) {
+            return " ESCAPE ''";
         }
-        if (expression.charAt(0) == '\'') {
-            return expression.replace(defaultEscapeCharacter.toString(), defaultEscapeCharacter.toString() + defaultEscapeCharacter);
-        } else {
-            return jpaProvider.getCustomFunctionInvocation("REPLACE", 1)
-                    + expression + ", '" + defaultEscapeCharacter + defaultEscapeCharacter + "', '"
-                    + defaultEscapeCharacter + defaultEscapeCharacter + defaultEscapeCharacter + defaultEscapeCharacter + "')";
-        }
+        return "";
     }
 
     protected String renderNullPrecedence(String expression, String resolvedExpression, String order, String nulls) {